AppDir/usr/bin/requirements.txt
# Requirements für die AppImage
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
altair>=5.0.0
